split netlify function into two separate functions
