+++ projects/ui/src/functions/ebipdata/ebipList.ts
