+++ projects/ui/src/functions/ebipdata/ebipdata.ts
@@ -3,12 +3,10 @@ import * as fs from 'fs';
 import middy from 'middy';
 import path from 'path';
 import { cors, rateLimit } from '../middleware';
-import { oldBipList } from './oldBipList';
-import { oldBipVoteData } from './oldBipVoteData';
 import { ebipList } from './ebipList';
 
 /**
- * Return BIP content for prior on-chain BIPs.
+ * Return EBIP content.
  */
 const _handler: Handler = async (event) => {
   try {
@@ -46,50 +44,6 @@ const _handler: Handler = async (event) => {
           };
         };
     
-        return {
-          statusCode: 200,
-          headers: {
-            'Content-Type': 'application/json',
-          },
-          body: JSON.stringify(output),
-        };
-      };
-    };
-
-    if (event.queryStringParameters?.getOldBip) {
-      if (event.queryStringParameters?.getOldBip === 'all') {
-        return {
-          statusCode: 200,
-          headers: {
-            'Content-Type': 'application/json',
-          },
-          body: JSON.stringify(oldBipList),
-        };
-      };
-
-      const bipNumber = Number(event.queryStringParameters?.getOldBip);
-      if (bipNumber >= 0) {
-        const proposalBody = await new Promise((resolve, reject) => {
-          fs.readFile(path.join(__dirname, `./bips/bip-${bipNumber}.md`), 'utf8', (err, data) => {
-            if (err) {
-              return reject(err);
-            }
-            resolve(data);
-          });
-        });
-
-        const output = {
-          ...oldBipList[bipNumber],
-          body: proposalBody,
-          votes: oldBipVoteData[bipNumber]
-        };
-
-        if (!proposalBody) {
-          return {
-            statusCode: 404,
-          };
-        }
-
         return {
           statusCode: 200,
           headers: {
