+++ projects/ui/src/functions/ebipdata/ebips/ebip-0.md
