+++ projects/ui/src/functions/ebipdata/ebips/ebip-1.md
