+++ projects/ui/src/functions/ebipdata/ebips/ebip-10.md
