+++ projects/ui/src/functions/ebipdata/ebips/ebip-12.md
