+++ projects/ui/src/functions/ebipdata/ebips/ebip-2.md
