+++ projects/ui/src/functions/ebipdata/ebips/ebip-3.md
