+++ projects/ui/src/functions/ebipdata/ebips/ebip-6.md
