+++ projects/ui/src/functions/ebipdata/ebips/ebip-7.md
