+++ projects/ui/src/functions/oldbipdata/bips/bip-10.md
