+++ projects/ui/src/functions/oldbipdata/bips/bip-12.md
