+++ projects/ui/src/functions/oldbipdata/bips/bip-17.md
