+++ projects/ui/src/functions/oldbipdata/bips/bip-3.md
