+++ projects/ui/src/hooks/beanstalk/useProposalBlockData.ts
@@ -148,7 +148,7 @@ export default function useProposalBlockData(
     (async () => {
       try {
         if (oldBip) {
-        const getOldBips = await fetch(`/.netlify/functions/proposal?getOldBip=${bipNumber}`)
+        const getOldBips = await fetch(`/.netlify/functions/oldbipdata?getOldBip=${bipNumber}`)
           .then((response) => response.json())
           setOldBipVotes(getOldBips.votes);
           setLoadingOldBipVotes(false);
